Some minor refactors to make things easier to read
