+++ src/main/java/com/terminalvelocitycabbage/templates/ecs/components/MaterialComponent.java
@@ -1,11 +1,13 @@
 package com.terminalvelocitycabbage.templates.ecs.components;
 
+import com.terminalvelocitycabbage.engine.client.renderer.materials.Atlas;
 import com.terminalvelocitycabbage.engine.ecs.Component;
 import com.terminalvelocitycabbage.engine.registry.Identifier;
 
 public class MaterialComponent implements Component {
 
     Identifier texture;
+    Atlas atlas;
 
     @Override
     public void setDefaults() {
@@ -18,5 +20,10 @@ public class MaterialComponent implements Component {
 
     public void setTexture(Identifier texture) {
         this.texture = texture;
+    }
+
+    public void setTexture(Identifier texture, Atlas atlas) {
+        this.texture = texture;
+        this.atlas = atlas;
     }
 }
